Try to compensate latency.
Rubberband signals latency to us for both channels independently. We take the
higher latency and report it to the host. To the delay of the channel with the
lower latency we add the difference of the two latencies.

Unfortunately this only works if the host supports latency compensation, which
is not the case for Ardour busses at the time of writing.